+++ demo/src/main/java/com/example/demo/controller/TWSEController.java
@@ -1,17 +1,15 @@
-package com.example.demo.controller;
+package com.example.demo.Controller;
 import java.io.IOException;
 
-import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
 
-import com.example.demo.Component.StockComponent.StockIdParam;
+import com.example.demo.Service.ResponseService;
 import com.example.demo.Service.TWSEService;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -28,13 +26,13 @@ public class TWSEController {
         @RequestParam("stock_id")
         @NotEmpty(message = "it can not be empty.")
         String stock_id) {
-        String stockUrl= "https://tw.stock.yahoo.com/_td-stock/api/resource/StockServices.revenues;period=year;symbol="+ stock_id;
+        String stockUrl= "https://tw.stock.yahoo.com/_td-stock/api/resource/StockServices.revenues;period=year;symbol=" + stock_id;
         try{
             twse= new TWSEService(stockUrl, stringRedisTemplate);
             return twse.getCompanyYearlyRevenue(stock_id);
         }catch(IOException io){
             io.printStackTrace();
-            return twse.responseError(io.toString());
+            return ResponseService.responseError("99999", io.toString());
         }
     }
 }
